+++ lib/convertor/jdl_convertor.js
@@ -9,7 +9,7 @@ const fs = require('fs'),
     dateFormatForLiquibase = require('../utils/format_utils').dateFormatForLiquibase,
     areEntitiesEqual = require('../utils/object_utils').areEntitiesEqual,
     RELATIONSHIP_TYPES = require('../core/jhipster/relationship_types').RELATIONSHIP_TYPES,
-    JDLParser = require('../dsl/jdl_parser');
+    JDLParser = require('../parser/jdl_parser');
 
 module.exports = {
   convertToJson: convertToJson
